refactor(stats): rework sequelize queries
LinkStatisticsRepository updates click stats by finding the instance,
then incrementing. This may result in larger transactions than needed,
given that the increment can be done in a single statement. Rework
the implementation so that we update, and if not present, we insert.
We will also do this in many small transactions.

Attempts to implement using `Model.upsert()` were not successful, due
to Sequelize creating temporary functions which fail to reference the
target table.

- interface - rework jsdoc, change return type to boolean
- impl - use `Model.update()` and `Sequelize.literal`, if not found,
  `Model.insert()` or throw not found error
- Drop the constraint of doing stats updates in a single txn